+++ coolamqp/clustering/cluster.py
@@ -71,10 +71,11 @@ class Cluster(object):
         self.extra_properties = extra_properties
         self.log_frames = log_frames
         self.on_blocked = on_blocked
+        self.connected = False
 
         if on_fail is not None:
             def decorated():
-                if not self.listener.terminating:
+                if not self.listener.terminating and self.connected:
                     on_fail()
 
             self.on_fail = decorated
@@ -221,7 +222,7 @@ class Cluster(object):
 
         self.snr = SingleNodeReconnector(self.node, self.attache_group,
                                          self.listener, self.extra_properties,
-                                         log_frames, self.name)
+                                         log_frames, self.name, self)
         self.snr.on_fail.add(lambda: self.events.put_nowait(ConnectionLost()))
         if self.on_fail is not None:
             self.snr.on_fail.add(self.on_fail)
@@ -230,7 +231,7 @@ class Cluster(object):
             self.snr.on_blocked.add(self.on_blocked)
 
         # Spawn a transactional publisher and a noack publisher
-        self.pub_tr = Publisher(Publisher.MODE_CNPUB)
+        self.pub_tr = Publisher(Publisher.MODE_CNPUB, cluster_to_set_connected_upon_first_connect=self)
         self.pub_na = Publisher(Publisher.MODE_NOACK)
         self.decl = Declarer()
 
@@ -245,9 +246,9 @@ class Cluster(object):
         if wait:
             # this is only going to take a short amount of time, so we're fine with polling
             start_at = monotonic.monotonic()
-            while not self.attache_group.is_online() and monotonic.monotonic() - start_at < timeout:
+            while not self.connected and monotonic.monotonic() - start_at < timeout:
                 time.sleep(0.1)
-            if not self.attache_group.is_online():
+            if not self.connected:
                 raise ConnectionDead(
                     '[%s] Could not connect within %s seconds' % (self.name, timeout,))